organize the Name module a bit
