+++ parser-typechecker/src/Unison/Builtin.hs
@@ -58,16 +58,16 @@ names = Names names0 mempty
 names0 :: Names0
 names0 = Names3.names0 terms types where
   terms = Rel.mapRan Referent.Ref (Rel.fromMap termNameRefs) <>
-    Rel.fromList [ (Name.fromVar vc, Referent.Con (R.DerivedId r) cid ct)
+    Rel.fromList [ (Name.unsafeFromVar vc, Referent.Con (R.DerivedId r) cid ct)
                  | (ct, (_,(r,decl))) <- ((CT.Data,) <$> builtinDataDecls @Symbol) <>
                     ((CT.Effect,) . (second . second) DD.toDataDecl <$> builtinEffectDecls)
                  , ((_,vc,_), cid) <- DD.constructors' decl `zip` [0..]] <>
-    Rel.fromList [ (Name.fromVar v, Referent.Ref (R.DerivedId i))
+    Rel.fromList [ (Name.unsafeFromVar v, Referent.Ref (R.DerivedId i))
                  | (v,i) <- Map.toList $ TD.builtinTermsRef @Symbol Intrinsic]
   types = Rel.fromList builtinTypes <>
-    Rel.fromList [ (Name.fromVar v, R.DerivedId r)
+    Rel.fromList [ (Name.unsafeFromVar v, R.DerivedId r)
                  | (v,(r,_)) <- builtinDataDecls @Symbol ] <>
-    Rel.fromList [ (Name.fromVar v, R.DerivedId r)
+    Rel.fromList [ (Name.unsafeFromVar v, R.DerivedId r)
                  | (v,(r,_)) <- builtinEffectDecls @Symbol ]
 
 -- note: this function is really for deciding whether `r` is a term or type,
