+++ parser-typechecker/src/Unison/Codebase/Editor/SlurpResult.hs
@@ -73,7 +73,7 @@ data SlurpResult v = SlurpResult {
 constructorsFor :: Var v => Set v -> UF.TypecheckedUnisonFile v Ann -> Set v
 constructorsFor types uf = let
   names = UF.typecheckedToNames0 uf
-  typesRefs = Set.unions $ Names.typesNamed names . Name.fromVar <$> toList types
+  typesRefs = Set.unions $ Names.typesNamed names . Name.unsafeFromVar <$> toList types
   ctorNames = R.filterRan isOkCtor (Names.terms names)
   isOkCtor (Referent.Con r _ _) | Set.member r typesRefs = True
   isOkCtor _ = False
