+++ parser-typechecker/src/Unison/FileParsers.hs
@@ -94,7 +94,7 @@ resolveNames typeLookupf preexistingNames uf = do
       possibleDeps = [ (Name.toText name, Var.name v, r) |
         (name, r) <- Rel.toList (Names.terms0 preexistingNames),
         v <- Set.toList (Term.freeVars tm),
-        name `Name.endsWithSegments` Name.fromVar v ]
+        name `Name.endsWithSegments` Name.unsafeFromVar v ]
       possibleRefs = Referent.toReference . view _3 <$> possibleDeps
   tl <- lift . lift . fmap (UF.declsToTypeLookup uf <>)
       $ typeLookupf (deps <> Set.fromList possibleRefs)
@@ -117,7 +117,7 @@ resolveNames typeLookupf preexistingNames uf = do
         [ (Var.name v, nr) |
           (name, r) <- Rel.toList (Names.terms0 $ UF.toNames uf),
           v <- Set.toList (Term.freeVars tm),
-          name `Name.endsWithSegments` Name.fromVar v,
+          name `Name.endsWithSegments` Name.unsafeFromVar v,
           typ <- toList $ TL.typeOfReferent tl r,
           let nr = Typechecker.NamedReference (Name.toText name) typ (Right r) ]
   pure (tm, fqnsByShortName, tl)
