+++ unison-core/src/Unison/DataDeclaration/Names.hs
@@ -32,10 +32,10 @@ toNames0 ct typeSymbol (Reference.DerivedId -> r) dd =
   -- constructor names
   foldMap names (DD.constructorVars dd `zip` [0 ..])
   -- name of the type itself
-  <> Names.names0 mempty (Rel.singleton (Name.fromVar typeSymbol) r)
+  <> Names.names0 mempty (Rel.singleton (Name.unsafeFromVar typeSymbol) r)
   where
   names (ctor, i) =
-    Names.names0 (Rel.singleton (Name.fromVar ctor) (Referent.Con r i ct)) mempty
+    Names.names0 (Rel.singleton (Name.unsafeFromVar ctor) (Referent.Con r i ct)) mempty
 
 dataDeclToNames :: Var v => v -> Reference.Id -> DataDeclaration v a -> Names0
 dataDeclToNames = toNames0 CT.Data
